requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.1.0
Flask-SocketIO>=5.3.0
Flask-Migrate>=4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
email-validator==2.1.0.post1
gunicorn>=21.0.0
redis>=5.0.0
eventlet>=0.33.0
psycopg2-binary>=2.9.0  # For PostgreSQL support
mysqlclient>=2.2.0
WTForms>=3.0.0
python-dateutil>=2.8.2
pytz>=2023.3
Pillow>=10.0.0
reportlab>=4.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
XlsxWriter>=3.1.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
mypy>=1.0.0
black>=23.0.0
types-python-dateutil>=2.8.0
types-pytz>=2023.3
types-redis>=4.6.0
sqlalchemy-stubs>=0.4